fixed failing tests and removed tests that install no longer valid code
