--- test/hotspot/jtreg/compiler/jvmci/events/JvmciNotifyInstallEventTest.config
@@ -1,2 +1,0 @@
-compiler.jvmci.events.JvmciNotifyInstallEventTest
-compiler.jvmci.common.JVMCIHelpers
