+++ test/hotspot/jtreg/compiler/jvmci/jdk.vm.ci.code.test/src/jdk/vm/ci/code/test/SimpleCodeInstallationTest.java
@@ -25,7 +25,7 @@
  * @test
  * @requires vm.jvmci
  * @requires vm.simpleArch == "x64" | vm.simpleArch == "aarch64" | vm.simpleArch == "riscv64"
- * @library /
+ * @library /test/lib /
  * @modules jdk.internal.vm.ci/jdk.vm.ci.hotspot
  *          jdk.internal.vm.ci/jdk.vm.ci.meta
  *          jdk.internal.vm.ci/jdk.vm.ci.code
@@ -39,8 +39,10 @@
  */
 
 package jdk.vm.ci.code.test;
+import jdk.test.lib.Asserts;
 
 import jdk.vm.ci.code.Register;
+import jdk.vm.ci.hotspot.HotSpotNmethod;
 import org.junit.Test;
 
 /**
@@ -61,6 +63,23 @@ public class SimpleCodeInstallationTest extends CodeInstallationTest {
 
     @Test
     public void test() {
-        test(SimpleCodeInstallationTest::compileAdd, getMethod("add", int.class, int.class), 5, 7);
+        HotSpotNmethod nmethod = test(SimpleCodeInstallationTest::compileAdd, getMethod("add", int.class, int.class), 5, 7);
+
+        // Test code invalidation
+        Asserts.assertTrue(nmethod.isValid(), "code is not valid, i = " + nmethod);
+        Asserts.assertTrue(nmethod.isAlive(), "code is not alive, i = " + nmethod);
+        Asserts.assertNotEquals(nmethod.getStart(), 0L);
+
+        // Make nmethod non-entrant but still alive
+        nmethod.invalidate(false);
+        Asserts.assertFalse(nmethod.isValid(), "code is valid, i = " + nmethod);
+        Asserts.assertTrue(nmethod.isAlive(), "code is not alive, i = " + nmethod);
+        Asserts.assertEquals(nmethod.getStart(), 0L);
+
+        // Deoptimize the nmethod and cut the link to it from the HotSpotNmethod
+        nmethod.invalidate(true);
+        Asserts.assertFalse(nmethod.isValid(), "code is valid, i = " + nmethod);
+        Asserts.assertFalse(nmethod.isAlive(), "code is alive, i = " + nmethod);
+        Asserts.assertEquals(nmethod.getStart(), 0L);
     }
 }
